ParsePDB.cpp: Change the way the correct line is selected and the coordinates are extracted.
It may very well be, that the coordinates are always placed at position 6,7,8, but there may be PDBs fir which that does not hold.
Therefore I used the columns (positions of the substr, basically) to get the correct informations.
The columns corresponding to the coordinates where provided by the PDB Format description, page 180 (176 document)